# reservation-infrastructure/src/main/resources/api-definition.yaml
---
openapi: 3.0.0
info:
  version: 1.0.0
  title: Reservation System API
  contact:
    email: paul@paulderaaij.nl
tags:
  - name: reservation-system
    description: Reserve tickets for events
paths:
  /events/available:
    get:
      summary: Retrieve available events for reservations
      tags:
        - available-events
      description: ''
      operationId: availableEvents
      responses:
        '200':
          description: returns a list of available events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailableEvent'
  /events/{id}/reservations:
    get:
      summary: Retrieve reservations for the requested event
      tags:
        - event-reservations
      description: ''
      operationId: eventReservations
      responses:
        200:
          description: returns a list of reservations placed for the event
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EventReservation'
    post:
      summary: Place a new reservation for the requested event
      tags:
        - make-reservation
      description: ''
      operationId: makeReservation
      responses:
        '201':
          description: successfully placed a reservation for the event
        '400':
          description: Not able to process the request due to invalid parameters
        '404':
          description: Event not found
      requestBody:
        $ref: '#/components/requestBodies/EventReservation'
components:
  requestBodies:
    EventReservation:
      type: object
      properties:
        eventId:
          type: string
        requestedTickets:
          type: integer
          format: int64
  schemas:
    AvailableEvent:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        capacity:
          type: integer
          format: int64
        date:
          type: string
          format: date
    EventReservation:
      type: object
      properties:
        id:
          type: string
        event:
          type: object
          $ref: '#/components/schemas/AvailableEvent'
        reservedTickets:
          type: integer
          format: int64